fix(doctor): change activity_id column type to unsigned integer
Align with the referenced activities table id column type to ensure foreign key constraint compatibility.

+++ packages/Webkul/Doctor/src/Database/Migrations/2026_01_19_120100_create_doctors_table.php
@@ -19,7 +19,7 @@ return new class extends Migration {
         Schema::create('doctor_activities', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('doctor_id');
-            $table->unsignedBigInteger('activity_id');
+            $table->unsignedInteger('activity_id');
             $table->foreign('doctor_id')->references('id')->on('doctors')->onDelete('cascade');
             $table->foreign('activity_id')->references('id')->on('activities')->onDelete('cascade');
         });
